updated dockerfile

+++ Commands.py
@@ -11,18 +11,19 @@ logger = logging.getLogger(__name__)
 
 # Get games from games.json
 async def get_games() -> list[str]:
-    games_list = []
-    try:
-        async with aiofiles.open(FILEPATH, "r") as f:
-            games_list = json.loads(await f.read())
-    except FileNotFoundError:
+    games_list: List[str] = []
+
+    if not FILEPATH.exists():
         async with aiofiles.open(FILEPATH, 'w') as f:
             await f.write(json.dumps(games_list, indent=4))
+    else:
+        async with aiofiles.open(FILEPATH, "r") as f:
+            games_list = json.loads(await f.read())
 
     return games_list
 
 
-async def get_rnd_game(author:str) -> str:
+async def get_rnd_game(author: str) -> str:
     games_list = await get_games()
     if not games_list:
         return "There is no game to choose from"
@@ -31,7 +32,7 @@ async def get_rnd_game(author:str) -> str:
     return rnd_game
 
 
-async def add_game(author:str, message:str) -> str:
+async def add_game(author: str, message: str) -> str:
     msg_list = message
     if msg_list == "":
         return "Enter a valid name"
@@ -45,7 +46,7 @@ async def add_game(author:str, message:str) -> str:
     return f"Added: {msg_list}"
 
 
-async def delete_game(author:str, message:str) -> str:
+async def delete_game(author: str, message: str) -> str:
     msg_list = message
     if msg_list == "":
         return "Enter a valid name"
@@ -64,7 +65,7 @@ async def delete_game(author:str, message:str) -> str:
         return f"Deleted: {msg_list}"
 
 
-async def reset_games(author:str) -> str:
+async def reset_games(author: str) -> str:
     async with aiofiles.open(FILEPATH, "w") as file:
         data = []
         await file.write(json.dumps(data, sort_keys=True, indent=4))
@@ -72,7 +73,7 @@ async def reset_games(author:str) -> str:
     return "Resetted games"
 
 
-async def list_games(author:str) -> str:
+async def list_games(author: str) -> str:
     games = await get_games()
     if not games:
         return "There is no game to choose from"
@@ -83,7 +84,7 @@ async def list_games(author:str) -> str:
 
 
 # CHANNELS
-async def clean(author:str, m:discord.message, client:discord.client) -> str:
+async def clean(author: str, m:discord.message, client:discord.client) -> str:
     temp = []
     async for msg in m.channel.history(limit=10000):
         if msg.author == client.user or msg.content.startswith('$'):
@@ -110,13 +111,11 @@ async def clean(author:str, m:discord.message, client:discord.client) -> str:
 
 
 # MISC
-async def status(author:str) -> str:
+async def status(author: str) -> str:
     uptime = psutil.boot_time()
     LoadCPU = psutil.cpu_percent()
     VMEM = psutil.virtual_memory()[2]
     logger.info(f"[{author}] Asked for status")
     return(
-        "Up since : "+uptime
-        + "\nCPU Load : "+LoadCPU
-        + '\nMemory % used: '+VMEM
+        f'Up since : {uptime}\nCPU Load : {LoadCPU}\nMemory % used: {VMEM}'
     )
